+++ slack-strava/commands/set.rb
@@ -4,16 +4,23 @@ module SlackStrava
       include SlackStrava::Commands::Mixins::Subscribe
 
       subscribe_command 'set' do |client, data, match|
+        user = ::User.find_create_or_update_by_slack_id!(client, data.user)
         if !match['expression']
           messages = [
             "Activities for team #{client.owner.name} display *#{client.owner.units_s}*.",
-            "Maps for team #{client.owner.name} are *#{client.owner.maps_s}*."
+            "Maps for team #{client.owner.name} are *#{client.owner.maps_s}*.",
+            "Your private activities will #{user.private_activities? ? '' : 'not'} be posted."
           ]
           client.say(channel: data.channel, text: messages.join("\n"))
-          logger.info "SET: #{client.owner} - set"
+          logger.info "SET: #{client.owner}, user=#{data.user} - set"
         else
           k, v = match['expression'].split(/\W+/, 2)
           case k
+          when 'private' then
+            changed = v && user.private_activities != v
+            user.update_attributes!(private_activities: v) unless v.nil?
+            client.say(channel: data.channel, text: "Your private activities will#{changed ? (user.private_activities? ? ' now' : ' no longer') : (user.private_activities? ? '' : ' not')} be posted.")
+            logger.info "SET: #{client.owner}, user=#{data.user} - private set to #{user.private_activities}"
           when 'units' then
             changed = v && client.owner.units != v
             client.owner.update_attributes!(units: v) unless v.nil?
@@ -25,7 +32,7 @@ module SlackStrava
             client.say(channel: data.channel, text: "Maps for team #{client.owner.name} are#{changed ? ' now' : ''} *#{client.owner.maps_s}*.")
             logger.info "SET: #{client.owner} - maps set to #{client.owner.maps}"
           else
-            raise "Invalid setting #{k}, you can _set units km|mi_ or _set maps off|image|thumb_."
+            raise "Invalid setting #{k}, type `help` for instructions."
           end
         end
       end