+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changelog
 
+* 2018/5/7: Don't sync private activities - [@dblock](https://github.com/dblock).
 * 2018/3/28: Initial public release - [@dblock](https://github.com/dblock).